fix: 에러 수정, 설정 파일 수정, schema.prisma 최신화

+++ Back/src/main/java/com/petmily/backend/api/walker/service/WalkerBookingService.java
@@ -59,7 +59,7 @@ public class WalkerBookingService {
             throw new CustomException(ErrorCode.INVALID_REQUEST, "Walker is not currently available");
         }
 
-        double hourlyRate = walker.getHourlyRate() != null ? walker.getHourlyRate() : 20000.0;
+        double hourlyRate = walker.getHourlyRate() != null ? walker.getHourlyRate().doubleValue() : 20000.0;
         double totalPrice = hourlyRate * (request.getDuration() / 60.0);
 
         WalkerBooking booking = new WalkerBooking();
